backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
pydantic==2.7.1
python-dotenv==1.0.1
google-generativeai==0.7.0
httpx==0.27.0
pyyaml==6.0.1
slowapi==0.1.9
python-multipart==0.0.9
